server/wenet/requirements.txt
absl-py==0.12.0
cachetools==4.2.1
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
chardet==4.0.0
flake8==3.8.2
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
Markdown==3.3.4
mccabe==0.6.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
Pillow @ file:///tmp/build/80754af9/pillow_1615224089586/work
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes==2.2.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sentencepiece==0.1.95
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
torch==1.6.0
torchaudio==0.6.0
torchvision==0.7.0
typeguard==2.11.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3==1.26.4
Werkzeug==1.0.1